% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confIntSD.R
\name{confIntSD}
\alias{confIntSD}
\title{Confidence interval for standard deviation}
\usage{
confIntSD(x, n = NULL, conf.level = 0.95)
}
\arguments{
\item{x}{Either a standard deviation, in which case \code{n} must also be
provided, or a vector, in which case \code{n} must be NULL.}

\item{n}{The sample size is \code{x} is a standard deviation.}

\item{conf.level}{The confidence level}
}
\value{
A vector or matrix.
}
\description{
This function is vectorized.
}
\examples{
rosetta::confIntSD(mtcars$mpg);
rosetta::confIntSD(c(6, 7), c(32, 32));
}
