% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatR.R
\name{formatR}
\alias{formatR}
\title{Pretty formatting of correlation coefficients}
\usage{
formatR(r, digits = 2)
}
\arguments{
\item{r}{The Pearson correlation to format.}

\item{digits}{The number of digits to round to.}
}
\value{
The formatted correlation.
}
\description{
Pretty formatting of correlation coefficients
}
\examples{
formatR(cor(mtcars$mpg, mtcars$disp));
}
\seealso{
\code{\link[=noZero]{noZero()}}, \code{\link[=formatCI]{formatCI()}}, \code{\link[=formatPvalue]{formatPvalue()}}
}
