% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqjmv.h.R
\name{freqjmv}
\alias{freqjmv}
\title{Frequencies}
\usage{
freqjmv(data, vector)
}
\arguments{
\item{data}{.}

\item{vector}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$table} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$table$asDF}

\code{as.data.frame(results$table)}
}
\description{
Frequencies
}
