% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pp15.R
\docType{data}
\name{partypanelData}
\alias{partypanelData}
\alias{pp15}
\title{Subsets of the Party Panel 2015 dataset}
\format{
A \code{data.frame} with 128 columns and 829 rows.
Note that many rows contain missing values; the columns and rows
were taken directly from the original Party Panel dataset, and
represent all participants that made it past a given behavior.
}
\usage{
data(pp15)
}
\description{
This is a subsets of the Party Panel 2015 dataset. Party Panel is an annual
semi-panel determinant study among Dutch nightlife patrons, where every
year, the determinants of another nightlife-related risk behavior are
mapped. In 2015, determinants were measured of behaviors related to using
highly dosed ecstasy pills.
}
\details{
The full dataset is publicly available through the Open Science
Framework (https://osf.io/s4fmu/). Also see the GitLab
repository (https://gitlab.com/partypanel) and the website
at https://partypanel.eu.
}
\examples{
data('pp15', package='rosetta');
rosetta::freq(pp15$gender);
}
\keyword{data}
