% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIMM.R
\name{plotIMM}
\alias{plotIMM}
\title{Makes plot of Index of Moderated Mediation of gemm object}
\usage{
plotIMM(x, ...)
}
\arguments{
\item{x}{object moderatedMediationSem}

\item{...}{optional}
}
\value{
simple slope plots for each mediator and simple slopes parameter estimates
}
\description{
Makes plot of Index of Moderated Mediation of gemm object
}
