% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSS.R
\name{plotSS}
\alias{plotSS}
\title{Makes simple slope plots of gemm object}
\usage{
plotSS(x, ...)
}
\arguments{
\item{x}{object moderatedMediationSem}

\item{...}{optional}
}
\value{
simple slope plots for each mediator and simple slopes parameter estimates
}
\description{
Makes simple slope plots of gemm object
}
