% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIMM3d.R
\name{prepIMM3d}
\alias{prepIMM3d}
\title{Computes Index of moderated mediation of gemm object}
\usage{
prepIMM3d(M1, M2, parEst = parEst, i = 1)
}
\arguments{
\item{M1}{moderator of x-m path}

\item{M2}{moderator of m-y path}

\item{parEst}{parameter estimates from lavaan results}

\item{i}{index of vector of mediators names}
}
\value{
vector of index of moderated mediation with CI limits for a given mediator
}
\description{
Computes Index of moderated mediation of gemm object
}
