% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomSlug.R
\name{randomSlug}
\alias{randomSlug}
\alias{idSlug}
\title{Generate a random slug}
\usage{
randomSlug(x = 10, id = NULL, chars = c(letters, LETTERS, 0:9))

idSlug(id = NULL, x = 10, chars = c(letters, LETTERS, 0:9))
}
\arguments{
\item{x}{Length of slug}

\item{id}{If not NULL, prepended to slug (separated with a dash) as id; in
that case, it's also braces and a hash is added.}

\item{chars}{Characters to sample from}
}
\value{
A character value.
}
\description{
idSlug is a convenience function with swapped argument order.
}
\examples{
randomSlug();
idSlug("identifier");
}
