% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeatStr.R
\name{repeatStr}
\alias{repeatStr}
\alias{repStr}
\title{Repeat a string a number of times}
\usage{
repeatStr(n = 1, str = " ")
}
\arguments{
\item{n, str}{Normally, respectively the frequency with which to
repeat the string and the string to repeat; but the order of the
inputs can be switched as well.}
}
\value{
A character vector of length 1.
}
\description{
Repeat a string a number of times
}
\examples{
### 10 spaces:
repStr(10);

### Three euro symbols:
repStr("\u20ac", 3);
}
