% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqjmv.h.R
\name{freqjmv}
\alias{freqjmv}
\title{Frequencies}
\usage{
freqjmv(data, vector)
}
\arguments{
\item{data}{.}

\item{vector}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Frequencies
}
