% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram}
\alias{histogram}
\title{Simple function to create a histogram}
\usage{
histogram(
  vector,
  bins = NULL,
  theme = ggplot2::theme_bw(),
  xLabel = NULL,
  yLabel = "Count"
)
}
\arguments{
\item{vector}{A variable or vector.}

\item{bins}{The number of bins; when 0, either the number of unique
values in \code{vector} or \code{20}, whichever is lower.}

\item{theme}{The ggplot2 theme to use.}

\item{xLabel, yLabel}{Labels for x and y axes; variable name is used for
x axis if no label is specified.}
}
\value{
A ggplot2 plot.
}
\description{
Simple function to create a histogram
}
\examples{
rosetta::histogram(mtcars$mpg);
}
